Move party_toast to new home
