+++ include/boe/adventurers.hpp
@@ -11,3 +11,5 @@ void adventurers_restore_sp(Adventurers& adventurers, short amt);
 bool adventurers_has_ability(const Adventurers& adventurers, short abil);
 short adventurers_luck_total(const Adventurers& adventurers);
 short adventurers_level_total(const Adventurers& adventurers);
+
+bool party_toast(const Adventurers& adventurers);
