+++ src/Windows Code Release 3/graphics.h
@@ -52,7 +52,6 @@ void redraw_terrain();
 void draw_targets(location center);
 void erase_spot(short i,short j);
 void draw_targeting_line(POINT where_curs);
-bool party_toast(const Adventurers& adventurers);
 void redraw_partial_terrain(RECT redraw_rect);
 void dump_gworld();
 void final_process_dialog(short which_dlog);
